change float to double in ex1-1.cpp.

+++ chapter01/ex1-1.cpp
@@ -7,7 +7,7 @@
 int main(void){
     int a,b,c;
     scanf("%d%d%d", &a, &b, &c);
-    float ave = (a + b + c)/3;
+    double ave = (a + b + c)/3;
     printf("%.3f\n", ave);
     return 0;
 }
